Diamond Dynamics Price Prediction and Market Segmentation/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
joblib
pyyaml
jupyter
notebook
ipykernel
xgboost
lightgbm
streamlit
pytest
python-dotenv